Test disabled stream wrappers in AssetFetcher
