+++ src/AssetFetcher.php
@@ -2,7 +2,6 @@
 
 namespace Mpdf;
 
-use Mpdf\Exception\AssetFetchingException;
 use Mpdf\File\LocalContentLoaderInterface;
 use Mpdf\File\StreamWrapperChecker;
 use Mpdf\Http\ClientInterface;
@@ -39,11 +38,11 @@ class AssetFetcher implements \Psr\Log\LoggerAwareInterface
 		$wrapperChecker = new StreamWrapperChecker($this->mpdf);
 
 		if ($wrapperChecker->hasBlacklistedStreamWrapper($path)) {
-			throw new AssetFetchingException('File contains an invalid stream. Only ' . implode(', ', $wrapperChecker->getWhitelistedStreamWrappers()) . ' streams are allowed.');
+			throw new \Mpdf\Exception\AssetFetchingException('File contains an invalid stream. Only ' . implode(', ', $wrapperChecker->getWhitelistedStreamWrappers()) . ' streams are allowed.');
 		}
 
 		if ($originalSrc && $wrapperChecker->hasBlacklistedStreamWrapper($originalSrc)) {
-			throw new AssetFetchingException('File contains an invalid stream. Only ' . implode(', ', $wrapperChecker->getWhitelistedStreamWrappers()) . ' streams are allowed.');
+			throw new \Mpdf\Exception\AssetFetchingException('File contains an invalid stream. Only ' . implode(', ', $wrapperChecker->getWhitelistedStreamWrappers()) . ' streams are allowed.');
 		}
 
 		$this->mpdf->GetFullPath($path);
